Check if the user requesting a session is valid
There are some cases when a user requesting a session shouldn't be
allowed to launch one, such as when the user is logged in as root, or
the user is a "functional account" (for example, p99user).

Check that the user satisfies these checks before launching a Hebi
session for them.

Note that the check of if a user is a visit user is not yet implemented.