Use file browser server for verifying session-readiness before redirection
In some cases of a newly created Hebi session, the file browser server
was not yet running but the nginx server was, which resulted in
redirecting a user to their session when it was not yet ready. Instead,
using an endpoint in the file browser server as a check for if the newly
created Hebi session is "ready" appeared to be more reliable.