Update Hebi manifest to work with custom Kyverno policy
In the upcoming shutdown at DLS starting on 21/10/22, a custom Kyverno
policy for the Hebi namespace will be deployed on Argus (the production
Kubernetes cluster at DLS).

Testing the current Hebi deployment against this custom Kyverno policy
in advance of the shutdown was possible via deploying on Pollux (a
development Kubernetes cluster at DLS), which had this custom Kyverno
policy applied already.

The testing revealed that setting
`containers[*].securityContext.privileged = false` in the Hebi template
Deployment manifest was causing an error with the custom Kyverno policy,
so that field has been removed.