Write session activity timestamps to file for persistence over launcher Pod restarts
When the launcher Pod was restarted, the timestamps of when all the
user's Hebi sessions were lost and thus it wasn't possible to determine
if a Hebi session was within the "session activity period" or not.

Periodically writing this information to a small file (whose storage is
provided by a Kubernetes PVC) and then loading this file if it exists
when the launcher container restarts allows this session timestamp
information to persist across launcher Pod restarts.

Also, there was a bug in the `check_if_pod_is_active()` function which
was causing the launcher to not detect inactive sessions correctly. This
was due to a mistake in assuming that the `difference.seconds` value
contained all the time information (ie, that the number of days in the
difference was converted into seconds and included in
`difference.seconds`). This has been fixed and now the `difference.days`
attribute is now taken into account in the check of whether a session is
active or not.